Move migration example to AppDelegate.
