+++ examples/ios/swift/Migration/AppDelegate.swift
@@ -17,6 +17,7 @@
 ////////////////////////////////////////////////////////////////////////////
 
 import UIKit
+import RealmSwift
 
 @UIApplicationMain
 class AppDelegate: UIResponder, UIApplicationDelegate {
@@ -29,12 +30,48 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
         window?.makeKeyAndVisible()
 
         #if CREATE_EXAMPLES
-        MigrationExample().addExampleDataToRealm(exampleData)
+        addExampleDataToRealm(exampleData)
         #else
-        MigrationExample().performMigration()
+        performMigration()
         #endif
 
         return true
+    }
+    
+    private func addExampleDataToRealm(_ exampleData: (Realm) -> Void) {
+        let url = realmUrl(for: schemaVersion, usingTemplate: false)
+        let configuration = Realm.Configuration(fileURL: url, schemaVersion: UInt64(schemaVersion))
+        let realm = try! Realm(configuration: configuration)
+
+        try! realm.write {
+            exampleData(realm)
+        }
+    }
+
+    private func performMigration() {
+        for realmVersion in 0..<schemaVersion {
+            let url = realmUrl(for: realmVersion, usingTemplate: true)
+            let realmConfiguration = Realm.Configuration(fileURL: url, schemaVersion: UInt64(schemaVersion), migrationBlock: migrationBlock)
+            try! Realm.performMigration(for: realmConfiguration)
+            let realm = try! Realm(configuration: realmConfiguration)
+            migrationCheck(realm)
+        }
+    }
+
+    private func realmUrl(for schemaVersion: Int, usingTemplate: Bool) -> URL {
+        let defaultURL = Realm.Configuration.defaultConfiguration.fileURL!
+        let defaultParentURL = defaultURL.deletingLastPathComponent()
+        let fileName = "default-v\(schemaVersion)"
+        let destinationUrl = defaultParentURL.appendingPathComponent(fileName + ".realm")
+        if FileManager.default.fileExists(atPath: destinationUrl.path) {
+            try! FileManager.default.removeItem(at: destinationUrl)
+        }
+        if usingTemplate {
+            let bundleUrl = Bundle.main.url(forResource: fileName, withExtension: "realm")!
+            try! FileManager.default.copyItem(at: bundleUrl, to: destinationUrl)
+        }
+
+        return destinationUrl
     }
 
 }
